code analysis done.

+++ src/main/java/inventory/controller/MainScreenController.java
@@ -1,4 +1,3 @@
-
 package inventory.controller;
 
 import inventory.model.Part;
@@ -15,30 +14,20 @@ import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
 import java.net.URL;
 import java.util.Optional;
 import java.util.ResourceBundle;
 
-public class MainScreenController implements Initializable,Controller {
-    
-     // Declare fields
-    private Stage stage;
-    private Parent scene;
-    private static Part modifyPart;
-    private static Product modifyProduct;
-    private static int modifyPartIndex;
-    private static int modifyProductIndex;
-    
-    // Declare methods
-    public static int getModifyPartIndex() {
-        return modifyPartIndex;
-    }
-    
-    public static int getModifyProductIndex() {
-        return modifyProductIndex;
-    }
+public class MainScreenController implements Initializable, Controller {
+    // Declare fields
+    private final Logger logger = LogManager.getLogger(MainScreenController.class);
+
+    private int modifyPartIndex;
+    private int modifyProductIndex;
 
     private InventoryService service;
 
@@ -72,21 +61,19 @@ public class MainScreenController implements Initializable,Controller {
 
     @FXML
     private TableColumn<Product, Double> productsPriceCol;
-    
+
     @FXML
     private TextField partsSearchTxt;
-    
+
     @FXML
     private TextField productsSearchTxt;
 
-    public MainScreenController(){}
-
-    public void setService(InventoryService service){
-        this.service=service;
+    // Declare methods
+    public void setService(InventoryService service) {
+        this.service = service;
         partsTableView.setItems(service.getAllParts());
         productsTableView.setItems(service.getAllProducts());
     }
-
 
     /**
      * Initializes the controller class and populate table views.
@@ -108,27 +95,36 @@ public class MainScreenController implements Initializable,Controller {
 
     /**
      * Method to add to button handler to switch to scene passed as source
-     * @param event
-     * @param source
-     * @throws IOException
+     *
+     * @param event  The event that happened
+     * @param source The source of the scene
+     * @throws IOException If there is an error in displaying the scene
      */
     private void displayScene(ActionEvent event, String source) throws IOException {
-        stage = (Stage)((Button)event.getSource()).getScene().getWindow();
-        FXMLLoader loader= new FXMLLoader(getClass().getResource(source));
-        //scene = FXMLLoader.load(getClass().getResource(source));
-        scene = loader.load();
-        Controller ctrl=loader.getController();
+        // Declare fields
+        Stage stage = (Stage) ((Button) event.getSource()).getScene().getWindow();
+
+        FXMLLoader loader = new FXMLLoader(getClass().getResource(source));
+        Parent scene = loader.load();
+
+        Controller ctrl = loader.getController();
         ctrl.setService(service);
+
+        if (ctrl instanceof ModifyProductController modifyProductController) {
+            modifyProductController.setProductIndex(modifyProductIndex);
+        } else if (ctrl instanceof ModifyPartController modifyPartController) {
+            modifyPartController.setPartIndex(modifyPartIndex);
+        }
+
         stage.setScene(new Scene(scene));
         stage.show();
     }
 
     /**
      * Ask user for confirmation before deleting selected part from list of parts.
-     * @param event 
      */
     @FXML
-    void handleDeletePart(ActionEvent event) {
+    void handleDeletePart() {
         Part part = partsTableView.getSelectionModel().getSelectedItem();
 
         Alert alert = new Alert(AlertType.CONFIRMATION);
@@ -138,41 +134,41 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
-            System.out.println("Part deleted.");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            logger.info("Part deleted.");
             service.deletePart(part);
         } else {
-            System.out.println("Canceled part deletion.");
+            logger.info("Canceled part deletion.");
         }
     }
 
     /**
      * Ask user for confirmation before deleting selected product from list of products.
-     * @param event 
      */
     @FXML
-    void handleDeleteProduct(ActionEvent event) {
+    void handleDeleteProduct() {
         Product product = productsTableView.getSelectionModel().getSelectedItem();
-        
+
         Alert alert = new Alert(AlertType.CONFIRMATION);
         alert.initModality(Modality.NONE);
         alert.setTitle("Confirmation");
         alert.setHeaderText("Confirm Product Deletion?");
         alert.setContentText("Are you sure you want to delete product " + product.getName() + " from products?");
         Optional<ButtonType> result = alert.showAndWait();
-        
-        if (result.get() == ButtonType.OK) {
+
+        if (result.isPresent() && result.get() == ButtonType.OK) {
             service.deleteProduct(product);
-            System.out.println("Product " + product.getName() + " was removed.");
+            logger.info("Product " + product.getName() + " was removed.");
         } else {
-            System.out.println("Product " + product.getName() + " was not removed.");
+            logger.info("Product " + product.getName() + " was not removed.");
         }
     }
 
     /**
      * Switch scene to Add Part
-     * @param event
-     * @throws IOException
+     *
+     * @param event The event that happened
+     * @throws IOException If there is an error loading the FXML file
      */
     @FXML
     void handleAddPart(ActionEvent event) throws IOException {
@@ -181,8 +177,9 @@ public class MainScreenController implements Initializable,Controller {
 
     /**
      * Switch scene to Add Product
-     * @param event
-     * @throws IOException
+     *
+     * @param event The event that happened
+     * @throws IOException If there is an error loading the FXML file
      */
     @FXML
     void handleAddProduct(ActionEvent event) throws IOException {
@@ -192,66 +189,65 @@ public class MainScreenController implements Initializable,Controller {
     /**
      * Changes scene to Modify Part screen and passes values of selected part
      * and its index
-     * @param event
-     * @throws IOException
+     *
+     * @param event The event that happened
+     * @throws IOException If there is an error loading the FXML file
      */
     @FXML
     void handleModifyPart(ActionEvent event) throws IOException {
-        modifyPart = partsTableView.getSelectionModel().getSelectedItem();
+        Part modifyPart = partsTableView.getSelectionModel().getSelectedItem();
         modifyPartIndex = service.getAllParts().indexOf(modifyPart);
-        
+
         displayScene(event, "/fxml/ModifyPart.fxml");
     }
 
     /**
      * Switch scene to Modify Product
-     * @param event
-     * @throws IOException
+     *
+     * @param event The event that happened
+     * @throws IOException If there is an error loading the FXML file
      */
     @FXML
     void handleModifyProduct(ActionEvent event) throws IOException {
-        modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
+        Product modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
         modifyProductIndex = service.getAllProducts().indexOf(modifyProduct);
-        
+
         displayScene(event, "/fxml/ModifyProduct.fxml");
     }
 
     /**
      * Ask user for confirmation before exiting
-     * @param event 
      */
     @FXML
-    void handleExit(ActionEvent event) {
+    void handleExit() {
         Alert alert = new Alert(AlertType.CONFIRMATION);
         alert.initModality(Modality.NONE);
         alert.setTitle("Confirmation Needed");
         alert.setHeaderText("Confirm Exit");
         alert.setContentText("Are you sure you want to exit?");
         Optional<ButtonType> result = alert.showAndWait();
-        if(result.get() == ButtonType.OK) {
-            System.out.println("Ok selected. Program exited");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            logger.info("OK clicked, exiting program.");
             System.exit(0);
         } else {
-            System.out.println("Cancel clicked.");
+            logger.info("Cancel clicked, returning to main screen.");
         }
     }
 
     /**
      * Gets search text and inputs into lookupPart method to highlight desired part
-     * @param event 
      */
     @FXML
-    void handlePartsSearchBtn(ActionEvent event) {
+    void handlePartsSearchBtn() {
         String x = partsSearchTxt.getText();
         partsTableView.getSelectionModel().select(service.lookupPart(x));
     }
 
     /**
      * Gets search text and inputs into lookupProduct method to highlight desired product
-     * @param event 
      */
     @FXML
-    void handleProductsSearchBtn(ActionEvent event) {
+    void handleProductsSearchBtn() {
         String x = productsSearchTxt.getText();
         productsTableView.getSelectionModel().select(service.lookupProduct(x));
     }
